Calculating headways in 30 minute buckets
