+++ mbta-performance/.chalice/config.json
@@ -21,7 +21,8 @@
         "process_daily_lamp": {
           "iam_policy_file": "policy-lamp-ingest.json",
           "lambda_timeout": 900,
-          "lambda_memory_size": 2048
+          "lambda_memory_size": 4096,
+          "max_ebs_size_gb": 10
         }
       }
     }
