+++ mbta-performance/app.py
@@ -9,7 +9,7 @@ app = Chalice(app_name="mbta-performance")
 app.register_middleware(ConvertToMiddleware(datadog_lambda_wrapper))
 
 
-# Runs every 60 minutes from either 4 AM -> 1:55AM or 5 AM -> 2:55 AM depending on DST
-@app.schedule(Cron("0", "0-6,9-23", "*", "*", "?", "*"))
+# Runs every 30 minutes from either 4 AM -> 1:55AM or 5 AM -> 2:55 AM depending on DST
+@app.schedule(Cron("*/30", "0-6,9-23", "*", "*", "?", "*"))
 def process_daily_lamp(event):
     lamp.ingest_today_lamp_data()
